fix click listeners not being deregistered
The code from the StackOverflow answer was buggy, because the DOM nodes
were being coerced into string when used as object keys.

This caused all event listeners to be registered as if they were
attached to the same document, which made the listener deregistering
working only for the first SVG element.

This fix the problem by storing the node reference in an object and
doing a sequential search to find the matching handlers to remove.